requirements.txt
Django==4.0.2
django_heroku==0.3.1
djangorestframework==3.13.1
itsdangerous==2.1.0
psycopg2-binary==2.9.3
gunicorn==20.1.0